Enhance number of functions for USB Host
With multiple USB host ports on boards like Adafruit Fruit Jam, provide a robust library of functions for checking what kind of devices are attached. This is important when a game or program wants to check the correct hardware is enumerating on the USB bus. Two tests, a simpletest and a more thorough test are in examples.